@@ -1,6 +1,6 @@
 {
 	"name": "skeleton-share",
-	"version": "1.0.0",
+	"version": "2.0.0",
 	"license": "MIT",
 	"scripts": {
 		"dev": "vite dev",
@@ -25,7 +25,9 @@
 	},
 	"typesVersions": {
 		">4.0": {
-			"providers": ["./dist/providers/index.d.ts"]
+			"providers": [
+				"./dist/providers/index.d.ts"
+			]
 		}
 	},
 	"files": [
@@ -34,40 +36,43 @@
 		"!dist/**/*.spec.*"
 	],
 	"peerDependencies": {
-		"svelte": "^4.0.0",
-		"tailwindcss": "^3.3.6"
+		"svelte": "^5.20.0",
+		"tailwindcss": "^4.0.9"
 	},
 	"devDependencies": {
+		"@floating-ui/dom": "^1.6.3",
 		"@iconify/svelte": "^4.0.0-beta.4",
-		"@skeletonlabs/skeleton": "^2.8.0",
-		"@skeletonlabs/tw-plugin": "^0.3.1",
-		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/package": "^2.2.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
-		"@types/eslint": "8.56.0",
-		"@types/node": "^20.11.19",
-		"@typescript-eslint/eslint-plugin": "^6.0.0",
-		"@typescript-eslint/parser": "^6.0.0",
-		"autoprefixer": "^10.4.16",
-		"eslint": "^8.56.0",
-		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.35.1",
-		"postcss": "^8.4.32",
-		"postcss-load-config": "^5.0.2",
-		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
-		"prettier-plugin-tailwindcss": "^0.5.9",
-		"publint": "^0.1.9",
-		"svelte": "^4.2.7",
-		"svelte-check": "^3.6.0",
-		"tailwindcss": "^3.3.6",
+		"@skeletonlabs/skeleton": "^3.1.1",
+		"@skeletonlabs/skeleton-svelte": "^1.0.0",
+		"@sveltejs/adapter-auto": "^5.0.0",
+		"@sveltejs/kit": "^2.20.2",
+		"@sveltejs/package": "^2.3.10",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
+		"@types/eslint": "^9.6.1",
+		"@types/node": "^22.13.13",
+		"@typescript-eslint/eslint-plugin": "^8.28.0",
+		"@typescript-eslint/parser": "^8.28.0",
+		"autoprefixer": "^10.4.21",
+		"eslint": "^9.23.0",
+		"eslint-config-prettier": "^10.1.1",
+		"eslint-plugin-svelte": "^3.3.3",
+		"postcss-load-config": "^6.0.1",
+		"prettier": "^3.5.3",
+		"prettier-plugin-svelte": "^3.2.6",
+		"prettier-plugin-tailwindcss": "^0.6.11",
+		"publint": "^0.3.9",
+		"svelte": "^5.25.3",
+		"svelte-check": "^4.1.5",
+		"tailwindcss": "^4.0.17",
 		"tslib": "^2.4.1",
-		"typescript": "^5.0.0",
-		"vite": "^5.0.11",
-		"@floating-ui/dom": "^1.6.3"
+		"typescript": "^5.8.2",
+		"vite": "^6.2.3"
 	},
 	"svelte": "./dist/index.js",
 	"types": "./dist/index.d.ts",
-	"type": "module"
+	"type": "module",
+	"dependencies": {
+		"@skeletonlabs/tw-plugin": "^0.4.1",
+		"@tailwindcss/vite": "^4.0.17"
+	}
 }
