@@ -1,30 +1,24 @@
-import { type Drawer, type DrawerStore } from '@skeletonlabs/skeleton';
-import { get } from 'svelte/store';
-import { isShareDrawer as isShareDrawerBase} from '$lib/share/ShareDrawerSettings.js';
-import type { EventDispatcher } from 'svelte';
-import type { ShareEvents } from './share/ShareEvents.js';
+import type { ShareDrawerState } from './share/ShareDrawerState.svelte.js';
 const extraOffsetPercentage = 0.5;
 
-let drawerStore: DrawerStore | undefined = undefined;
 let initial = { x: 0, y: 0, height: 0 };
 let drawer: HTMLElement | null = null;
 let prev = 0;
 let momentum = 0;
 
-let dispatch: EventDispatcher<ShareEvents> | undefined = undefined;
+let onClose = $state<(() => void) | undefined>(undefined);
+let drawerState: ShareDrawerState | undefined;
 
 const isTouchEvent = (e: Event): e is TouchEvent => window.TouchEvent && e instanceof TouchEvent;
-const isShareDrawer = ()  => isShareDrawerBase(drawerStore);
 
 function setTranslate(drawer: HTMLElement, y: number) {
 	drawer.style.transform = `translateY(${y}px)`;
 }
 
-function handleMouseDown(e: CustomEvent<MouseEvent> | TouchEvent) {
-	if(!isShareDrawer()) return; 
-	if (e.detail instanceof MouseEvent && e.detail.button !== 0) return;
-	
-	if (!drawerStore || !get(drawerStore).open) return;
+function handleMouseDown(e: MouseEvent | TouchEvent) {
+	if (e instanceof MouseEvent && e.button !== 0) return;
+
+	//if (!drawerStore || !get(drawerStore).open) return;
 	const drawerContent = document.querySelector('.drawer-content');
 	if (
 		drawerContent &&
@@ -36,8 +30,8 @@ function handleMouseDown(e: CustomEvent<MouseEvent> | TouchEvent) {
 	if (!drawer) return;
 	const isTouch = isTouchEvent(e);
 	if (!isTouch) e.preventDefault();
-	const clientX = isTouch ? e.touches[0].clientX : e.detail.clientX;
-	const clientY = isTouch ? e.touches[0].clientY : e.detail.clientY;
+	const clientX = isTouch ? e.touches[0].clientX : e.clientX;
+	const clientY = isTouch ? e.touches[0].clientY : e.clientY;
 	initial = {
 		x: clientX,
 		y: clientY,
@@ -49,10 +43,8 @@ function handleMouseDown(e: CustomEvent<MouseEvent> | TouchEvent) {
 }
 
 function handleMouseMove(e: MouseEvent | TouchEvent) {
-	if(!isShareDrawer()) return;
 	if (!drawer) return;
 
-	if (!drawerStore || !get(drawerStore).open) return;
 	const clientY = isTouchEvent(e) ? e.touches[0].clientY : e.clientY;
 	let diff = initial.y - clientY;
 	if (initial.y - clientY < 0) {
@@ -77,9 +69,8 @@ function reset(transform: boolean) {
 }
 
 function handleMouseUp(e: MouseEvent | TouchEvent) {
-	if(!isShareDrawer()) return;
 	if (e instanceof MouseEvent && e.button !== 0) return;
-	
+
 	if (!drawer) return;
 	const target = e.target as HTMLElement;
 	if (target && target.classList !== undefined && target.classList.contains('drawer-backdrop')) {
@@ -95,52 +86,39 @@ function handleMouseUp(e: MouseEvent | TouchEvent) {
 		initial.height + diff < initial.height / 2 &&
 		(momentum < 0 || (initial.height + diff) * 0.1 < initial.height / 2)
 	) {
-		drawerStore && drawerStore.close();
-		if(dispatch) dispatch('close');
+		if (drawerState) drawerState.open = false;
+		onClose?.();
 		reset(false);
 	} else {
 		reset(true);
 	}
 }
 
-function handleBackdrop(e: CustomEvent<MouseEvent>) {
-	if(!isShareDrawer()) return;
-	if (!drawer) return;
-
-	e.preventDefault();
-	setTimeout(() => {
-		if (!drawer && drawerStore) drawerStore.close();
-		if(dispatch) dispatch('close');
-	}, 2);
-}
-
 function handleSelect(e: Event) {
 	if (!drawer) return;
 	e.preventDefault();
 }
-export function initDrawer(drawerComponent: Drawer | undefined, drawerStoreInstance: DrawerStore, eventDispatcher: EventDispatcher<ShareEvents>) {
-	if (!drawerComponent) return;
-	dispatch = eventDispatcher;
-	drawerStore = drawerStoreInstance;
-	drawerComponent.$on('drawer', handleMouseDown);
-	drawerComponent.$on('touchstart', handleMouseDown);
-	drawerComponent.$on('backdrop', handleBackdrop);
+export function initDrawer(newDrawerState: ShareDrawerState, onCloseCallback: typeof onClose) {
+	drawerState = newDrawerState;
+	onClose = onCloseCallback;
+
 	window.addEventListener('mousemove', handleMouseMove);
 	window.addEventListener('touchmove', handleMouseMove);
+	window.addEventListener('mousedown', handleMouseDown);
+	window.addEventListener('touchstart', handleMouseDown);
 	window.addEventListener('mouseup', handleMouseUp);
 	window.addEventListener('touchend', handleMouseUp);
 	window.addEventListener('selectstart', handleSelect);
 }
 
-export function destroyDrawer(drawerComponent: Drawer | undefined) {
-	if (!drawerComponent) return;
+export function destroyDrawer() {
+	if (!window) return;
 	window.removeEventListener('mousemove', handleMouseMove);
 	window.removeEventListener('touchmove', handleMouseMove);
 	window.removeEventListener('mouseup', handleMouseUp);
 	window.removeEventListener('touchend', handleMouseUp);
 	window.removeEventListener('selectstart', handleSelect);
 
-	drawerStore = undefined;
 	initial = { x: 0, y: 0, height: 0 };
 	drawer = null;
 	prev = 0;
