@@ -1,4 +1,8 @@
 // Reexport your entry components here
-export { isShareDrawer, shareDrawerSettings, type ShareDrawerData } from "./share/ShareDrawerSettings.js";
-export { share, type Share } from "./share.js";
-export {default as ShareDrawer} from "./share/ShareDrawer.svelte";
+export {
+	type ShareDrawerData,
+	openShareDrawer,
+	closeShareDrawer,
+} from './share/ShareDrawerState.svelte.js';
+export { share, type Share } from './share.js';
+export { default as ShareDrawer } from './share/ShareDrawer.svelte';
